Convert tolerances metaclass to a mixin class
